+++ index.js
@@ -37,55 +37,31 @@ var archi = arch
 const path = require("path")
 const fs = require("fs");
 const { error } = require("console");
-const package_root = path.join(process.cwd(), "package.json")
-const package_root_builder = path.resolve(".", "resources", "app", "package.json")
+function package_location (){
+    if (fs.existsSync(path.join(process.cwd(), "package.json"))) return path.join(process.cwd(), "package.json")
+    else if (fs.existsSync(path.resolve(".", "resources", "app", "package.json"))) return path.resolve(".", "resources", "app", "package.json")
+    else if (fs.existsSync(path.resolve(__dirname, "package.json"))) return path.resolve(__dirname, "package.json")
+    else path.resolve(".", "package.json")
+}
+const package_root = package_location();
 var cache_dir,home,desktop,tmp,system
+module.exports.package_path = package_root
+const tmp_men = "We were unable to locate some files, a temporary folder will be created, after restarting the server or computer some saved functions will be lost"
 if (process.platform == "win32") {
-<<<<<<< HEAD
-    var home = process.env.USERPROFILE;
-    var desktop = path.join(home, "Desktop")
-    if (fs.existsSync(package_root)){
-        var cache_dir = path.join(home, "AppData", "Roaming", require(package_root).name)
-    } else {
-        console.warn(`Temporary Storages, some functions will be lost after restarting the system`);
-        var cache_dir = path.join(process.env.TMP, `bds_tmp_configs`);
-    }
-    var tmp = process.env.TMP
-    var system = `windows`;
-} else if (process.platform == "linux") {
-    var home = process.env.HOME;
-    if (fs.existsSync(package_root)){
-        var cache_dir = path.join(home, ".config", require(package_root).name);
-    } else {
-        console.warn(`Temporary Storages, some functions will be lost after restarting the system`);
-        var cache_dir = `/tmp/bds_tmp_configs`;
-    }
-    var file = path.join(home, ".config", "user-dirs.dirs");var data = {};
-    if (fs.existsSync(file)){let content = fs.readFileSync(file,"utf8");let lines = content.split(/\r?\n/g).filter((a)=> !a.startsWith("#"));for(let line of lines){let i = line.indexOf("=");if(i >= 0){try{data[line.substring(0,i)] = JSON.parse(line.substring(i + 1))}catch(e){}}}};if(data["XDG_DESKTOP_DIR"]){var desktop = data["XDG_DESKTOP_DIR"];desktop = desktop.replace(/\$([A-Za-z\-\_]+)|\$\{([^\{^\}]+)\}/g, (_, a, b) => (process.env[a || b] || ""))}else{var desktop = "/tmp"}
-    
-    var tmp = `/tmp`;
-    var system = `linux`;
-=======
     home = process.env.USERPROFILE;
     desktop = path.join(home, "Desktop")
-    if (fs.existsSync(package_root)){
-        cache_dir = path.join(home, "AppData", "Roaming", require(package_root).name)
-    } else if (package_root_builder){
-        cache_dir = path.join(home, "AppData", "Roaming", require(package_root_builder).name)
-    } else {
-        console.warn("Temporary Storages, some functions will be lost after restarting the system");
+    if (fs.existsSync(package_root)) cache_dir = path.join(home, "AppData", "Roaming", require(package_root).name)
+    else {
+        console.warn(tmp_men);
         cache_dir = path.join(process.env.TMP, "bds_tmp_configs");
     }
     tmp = process.env.TMP
     system = "windows";
 } else if (process.platform == "linux") {
     home = process.env.HOME;
-    if (fs.existsSync(package_root)){
-        cache_dir = path.join(home, ".config", require(package_root).name);
-    } else if (fs.existsSync(package_root_builder)) {
-        cache_dir = path.join(home, ".config", require(package_root_builder).name);
-    } else {
-        console.warn("Temporary Storages, some functions will be lost after restarting the system");
+    if (fs.existsSync(package_root)) cache_dir = path.join(home, ".config", require(package_root).name);
+    else {
+        console.warn(tmp_men);
         cache_dir = "/tmp/bds_tmp_configs";
     }
     var file = path.join(home, ".config", "user-dirs.dirs");var data = {};
@@ -109,7 +85,6 @@ if (process.platform == "win32") {
     
     tmp = "/tmp";
     system = "linux";
->>>>>>> main
 } else if (process.platform == "darwin") {
     require("open")("https://github.com/Bds-Maneger/Bds_Maneger/wiki/systems-support#a-message-for-mac-os-users")
     console.error("Please use Windows or Linux MacOS Not yet supported")
@@ -117,53 +92,6 @@ if (process.platform == "win32") {
 } else {
     console.log(`Please use an operating system (OS) compatible with Minecraft Bedrock Server ${process.platform} is not supported`);
     process.exit(2021)
-<<<<<<< HEAD
-};
-// ---------
-// ---------
-var bds_dir = path.join(home, "bds_Server");
-var bds_dir_bedrock = path.join(bds_dir, 'bedrock');
-var bds_dir_java = path.join(bds_dir, 'java');
-var bds_dir_backup = path.join(bds_dir, 'backups');
-module.exports.backup_folder = bds_dir_backup
-
-if (!(fs.existsSync(bds_dir))){
-    console.log("Creating the bds directory")
-    fs.mkdirSync(bds_dir)
-    if (!(fs.existsSync(bds_dir))) shell.mkdir("-p", bds_dir);
-}
-
-// Configs
-const bds_config_file = path.join(bds_dir, "bds_config.json")
-if (fs.existsSync(bds_config_file)){
-    var bds_config = JSON.parse(fs.readFileSync(bds_config_file, "utf-8"))
-} else {
-    const _config = `{
-        "bds_platform": "bedrock",
-        "telegram_token": null
-    }`
-    var bds_config = JSON.parse(_config)
-    fs.writeFileSync(bds_config_file, _config)
-}
-module.exports.platform = bds_config.bds_platform
-console.log(`Running on the \"${bds_config.bds_platform}\" platform`)
-// Configs
-
-var log_dir = path.join(bds_dir, "log");
-var log_file = path.join(log_dir, `${date()}_${bds_config.bds_platform}_Bds_log.log`);
-var log_date = date();
-
-// ---------
-// ---------
-
-
-
-if (!(fs.existsSync(cache_dir))){
-    console.log(`Creating a folder for Storage in ${cache_dir}`);
-    fs.mkdirSync(cache_dir)
-    if (!(fs.existsSync(cache_dir))) shell.mkdir("-p", cache_dir);
-}
-=======
 }
 // ---------
 // ---------
@@ -215,7 +143,6 @@ if (!(fs.existsSync(cache_dir))){
     fs.mkdirSync(cache_dir)
     if (!(fs.existsSync(cache_dir))) shell.mkdir("-p", cache_dir);
 }
->>>>>>> main
 // e
 if (!(fs.existsSync(bds_dir_java))){
     console.log("Creating the bds directory to Java")
@@ -233,33 +160,6 @@ if (!(fs.existsSync(log_dir))){
         console.log(`Creating the bds log dir (${log_dir})`)
         fs.mkdirSync(log_dir)
         if (!(fs.existsSync(log_dir))) shell.mkdir("-p", log_dir)
-<<<<<<< HEAD
-    };
-};
-// e
-if (require("fs").existsSync(`${bds_dir}/telegram_token.txt`)){
-    module.exports.token = fs.readFileSync(`${bds_dir}/telegram_token.txt`, "utf8").replaceAll("\n", "");
-} else {
-    module.exports.token = undefined;
-}
-
-module.exports.telegram_token_save = (token) =>{
-    fs.writeFileSync(`${bds_dir}/telegram_token.txt`, token);
-    return "OK"
-}
-if (typeof fetch === "undefined"){global.fetch = require("node-fetch")}
-
-if (typeof localStorage === "undefined"){
-    var localStorageS = require("node-localstorage").LocalStorage;
-    global.localStorage = new localStorageS(`${cache_dir}/Local_Storage`);
-}
-
-// Java or Bedrock
-if (process.env.JAVA_ENABLE !== undefined){
-    localStorage.setItem('bds_edititon', 'java');
-}else{
-    localStorage.setItem('bds_edititon', 'bedrock');
-=======
     }
 }
 // e
@@ -317,7 +217,6 @@ if (require("fs").existsSync(path.join(bds_dir, "telegram_token.txt"))){
     } catch {
         throw new error("It was not possible to move the old telegram token file to the new bds maneger api file")
     }
->>>>>>> main
 }
 
 if (process.env.BDS_MONI == blanks){process.env.BDS_MONI = "false"}
@@ -330,15 +229,12 @@ fetch("https://raw.githubusercontent.com/Bds-Maneger/Raw_files/main/credentials.
      * backup credentials and an interesting design feature, plus privacy is important
      */
     module.exports.google_drive_credential = gd_cre
-<<<<<<< HEAD
-    module.exports.mcpe_file = require("./global/auth").mcpe
-=======
     /**
      * download the latest version of minecraft bedrock for android available, remember to use if you want ✌
      * 
      * you are taking responsibility for that
      */
-    module.exports.mcpe_file = require("./global/auth").mcpe
+    module.exports.mcpe_file = require("./scripts/auth").mcpe
     /**
      * perform a backup of the map, some resources are still under construction in the code more works
      * 
@@ -346,22 +242,15 @@ fetch("https://raw.githubusercontent.com/Bds-Maneger/Raw_files/main/credentials.
      * 
      * on the java platform the map selected in the server configuration will be backed up, any other map will have to change in the server settings to perform the backup
      */
->>>>>>> main
-    module.exports.drive_backup= require("./global/auth").drive_backup
+    module.exports.drive_backup= require("./scripts/auth").drive_backup
 });
 fetch("https://raw.githubusercontent.com/Bds-Maneger/Raw_files/main/Server.json").then(response => response.json()).then(rawOUT => {
     const versions = Object.getOwnPropertyNames(rawOUT.bedrock);
     for (let v in versions){
         var html = `${versions[v]}`;
-<<<<<<< HEAD
-        module.exports.version_select += `<option value=\"${html}\">${html}</option>\n`;
-        v++;
-    };
-=======
         module.exports.version_select += `<option value="${html}">${html}</option>\n`;
         v++;
     }
->>>>>>> main
     module.exports.bedrock_all_versions = Object.getOwnPropertyNames(rawOUT.bedrock);
     module.exports.java_all_versions = Object.getOwnPropertyNames(rawOUT.java);
     module.exports.bds_latest = rawOUT.bedrock_lateste;
@@ -372,16 +261,9 @@ fetch("https://raw.githubusercontent.com/Bds-Maneger/Raw_files/main/Server.json"
         if (typeof bds_api_start === "undefined"){
             require("./API/api")();
             require("./API/log")();
-<<<<<<< HEAD
-            require("./API/remote_access")();
-        }
-    } else {
-        console.warn(`The API via http is disabled, for more information, visit https://docs.srherobrine23.com/enable_bds_requests.html`)
-=======
         }
     } else {
         console.warn("The API via http is disabled, for more information, visit https://docs.srherobrine23.com/enable_bds_requests.html")
->>>>>>> main
     }
     module.exports.get_version = (type) => {
         if (type == "raw")
@@ -441,7 +323,7 @@ module.exports.telegram_token = JSON.parse(fs.readFileSync(path.join(bds_dir, "b
 
 
 // Global commands
-module.exports.telegram = require("./global/telegram_bot")
+module.exports.telegram = require("./scripts/telegram_bot")
 module.exports.change_platform = platform_update
 module.exports.token_register = () => {
     if (!(fs.existsSync(path.join(bds_dir, "bds_tokens.json")))){
@@ -471,23 +353,7 @@ module.exports.date = date
  * 
  * @example bds.command("say hello from Bds Maneger")
  */
-module.exports.command = require("./global/command").command
-<<<<<<< HEAD
-// module.exports.stop = require("./global/stop").Server_stop
-
-// New management method
-module.exports.start = require("./new_script/basic_server").start
-module.exports.stop = require("./new_script/basic_server").stop
-module.exports.backup = require("./new_script/backups").World_BAckup
-module.exports.detect = require("./new_script/detect")
-module.exports.bds_detect = require("./new_script/detect")
-module.exports.version_Download = require("./new_script/bds_download")
-module.exports.download = require("./new_script/bds_download")
-module.exports.kill = require("./new_script/kill_server")
-module.exports.config_example = require("./new_script/bds_settings").config_example
-module.exports.set_config = require("./new_script/bds_settings").config
-module.exports.get_config = require("./new_script/bds_settings").get_config
-=======
+module.exports.command = require("./scripts/command").command
 // New management method
 
 /**
@@ -559,4 +425,3 @@ module.exports.set_config = require("./scripts/bds_settings").config
  * takes the server settings in JSON format
  */
 module.exports.get_config = require("./scripts/bds_settings").get_config
->>>>>>> main