+++ scripts/check.js
@@ -6,11 +6,6 @@ module.exports.checkUser = (USERNAME) => {
     if (fs.existsSync(telegram_admin)) {
         var admins = fs.readFileSync(telegram_admin, "utf8");
     } else {
-<<<<<<< HEAD
-        var admins = `{"sh23_bot_not_config": {"allow": true}}`;
-        console.warn("All allowed")
-        console.log(`Create file in with name: ${require("../index").bds_dir}/telegram_admin.json`)
-=======
         const config = {
             "sh23_bot_not_config": {
                 "allow": true
@@ -18,17 +13,12 @@ module.exports.checkUser = (USERNAME) => {
         }
         fs.writeFileSync(telegram_admin, JSON.stringify(config))
         throw new console.error(`we just created the telegram authorization, edit before using: ${config}`);
->>>>>>> main
     }
     var adm = JSON.parse(admins);
     for(let check_ in adm){
         if (USERNAME == check_){
             return true
-<<<<<<< HEAD
-        } else if (index == "sh23_bot_not_config"){
-=======
         } else if (check_ == "sh23_bot_not_config"){
->>>>>>> main
             console.warn("Allow all")
             return true
         } check_++;