+++ CHANGELOG.md
@@ -1,5 +1,14 @@
-<<<<<<< HEAD
-=======
+# 1.5.5
+
+System monitor update:
+-   now we can get as much ram as is being used by java and bedrock servers
+
+The old scripts in the global folder have been moved to the scripts folder, and where all the new scripts are currently located.
+
+API update;
+-   We didn't have many updates to the script, just a touch up
+-   Some features are being planned to be added
+
 # 1.5.0
 
 All scripts have been rewritten some features are still under maintenance
@@ -8,7 +17,6 @@ The Docker image has returned to work in the new version.
 
 Some functions will be modified or removed from the next version in the Stable branch
 
->>>>>>> main
 # 1.4.2
 
 Docker test fix